MacOS build support

+++ rfsim/sources/Simulator.cpp
@@ -50,8 +50,8 @@ namespace rfsim {
         }
 
         mConfigManager = std::make_shared<ConfigManager>("config.json");
-        mWindowWidth = mConfigManager->GetWindowSize().x;
-        mWindowHeight = mConfigManager->GetWindowSize().y;
+        mWindowWidth = (int) mConfigManager->GetWindowSize().x;
+        mWindowHeight = (int) mConfigManager->GetWindowSize().y;
 
         // Setup main window and window manager
         mWindowManager = std::make_shared<WindowManager>();
